Don't echo %0
Trying to echo %0 breaks the batch file with a syntax error when weird
quotations are used. The following command breaks both %0 and %~0:

  g"itkdbx

The author of this commit has waived all copyright and
related or neighboring rights to the extent permitted by
law as described by the CC0 1.0 Universal Public Domain
Dedication. You should have received a copy of the full
dedication along with this commit, typically as a file
named <CC0-1.0.txt>. If not, it may be available at
<https://creativecommons.org/publicdomain/zero/1.0/>.